+++ R/transfer_series.R
@@ -4,14 +4,15 @@
 #'
 #' @param ws_from The workspace containing the additionnal series
 #' @param ws_to The workspace to add series to
-#' @param mp_to_name The name of the multiprocessing to transfer the series to
-#' @param mp_from_name The name of the multipricessing to transfer the series from
+#' @param name_mp_to The name of the multiprocessing to transfer the series to
+#' @param name_mp_from The name of the multipricessing to transfer the series from
 #' @param print_indications A boolean to print indications on the processing status (optional)
+#' @param selected_series The vector containing the series-to-update's names.
 #'  
-#' @details If `mp_to_name` and `mp_from_name` are unspecified, the update will be performed using 
+#' @details If `name_mp_to` and `name_mp_from` are unspecified, the update will be performed using 
 #' the workspaces' first SAProcessing. If only one SAProcessing name is specified but the workspace 
-#' it refers to is not (e.g. argument `SAP1` instead of `mp_from_name="SAP1"`), it will be 
-#' attributed by default to the first worskpace (`mp_to_name="SAP1"`). 
+#' it refers to is not (e.g. argument `SAP1` instead of `name_mp_from="SAP1"`), it will be 
+#' attributed by default to the first worskpace (`name_mp_to="SAP1"`). 
 #'
 #' @return the `workspace` ws_to augmented with series present in ws_from and not already in ws_to
 #' @examples \dontrun{transfer_series(ws_from, ws_to, "SAProcessing-1", "MP2", TRUE)}
@@ -20,16 +21,19 @@
 #'
 transfer_series <- function(
         ws_from, ws_to, 
-        mp_from_name = NULL, mp_to_name = NULL, 
+        selected_series, 
+        name_mp_from, name_mp_to, 
+        pos_mp_from, pos_mp_to, 
         print_indications = FALSE, 
-        create = TRUE) {
+        create_mp = TRUE, 
+        replace_series = TRUE) {
     
     # Verification of the parameters type
-    if (!inherits(ws_to, "workspace")) {
-        stop("The argument ws_to must be a workspace")
-    }
     if (!inherits(ws_from, "workspace")) {
         stop("The argument ws_from must be a workspace")
+    }
+    if (!inherits(ws_to, "workspace")) {
+        stop("The argument ws_to must be a workspace")
     }
     
     # Check that the workspaces aren't empty
@@ -44,52 +48,121 @@ transfer_series <- function(
     # "sap" refers to all SAPs of each workspace
     # "mp" refers to the specified "mp_name" SAP to update
     
-    saps_to <- RJDemetra::get_all_objects(ws_to)
-    names_saps_to <- names(saps_to)
-    
     saps_from <- RJDemetra::get_all_objects(ws_from)
     names_saps_from <- names(saps_from)
     
-    # if SAPs have been specified, verification that both workspaces contain the SAP
-    # and storage of its position in each workspace
-    
-    # if no SAP has been declared, the first of each workspaces will be used
-    pos_mp_to <- 1  
-    pos_mp_from <- 1
-    
-    # otherwise, we retrieve the SAP's position in each workspace 
-    if (!missing(mp_to_name)) {
-        pos_mp_to <- which(names_saps_to == mp_to_name)
+    # Check MP selection
+    # WS from
+    if (missing(name_mp_from) && missing(pos_mp_from)) {
+        pos_mp_from <- 1L
+        name_mp_from <- names_saps_from[1L]
+        print(paste0("No sap has been selected, the first one (", name_mp_from, ") will be used."))
+    } else if (missing(name_mp_from)) {
+        if (pos_mp_from <= RJDemetra::get_all_objects(ws_from) && pos_mp_from > 0L) {
+            name_mp_from <- names_saps_from[pos_mp_from]
+        } else {
+            warning(paste0("There is no SAP n°", pos_mp_from))
+            print("The program stops without transferring the series.")
+            return(FALSE)
+        }
+    } else if (missing(pos_mp_from)) {
+        if (name_mp_from %in% names_saps_from) {
+            pos_mo_from <- which(names_saps_from == name_mp_from)
+        } else {
+            warning(paste0("There is no SAP named ", name_mp_from, "in ws_from."))
+            print("The program stops without transferring the series.")
+            return(FALSE)
+        }
+    } else {
+        if (!name_mp_from %in% names_saps_from) {
+            warning(paste0("There is no SAP named ", name_mp_from, "in ws_from."))
+            print("The program stops without transferring the series.")
+            return(FALSE)
+        } else if (pos_mp_from > RJDemetra::get_all_objects(ws_from) || pos_mp_from <= 0L) {
+            warning(paste0("There is no SAP n°", pos_mp_from))
+            print("The program stops without transferring the series.")
+            return(FALSE)
+        } else if (names_saps_from[pos_mp_from] != name_mp_from) {
+            warning("The arguments pos_mp_from and name_mp_from are refering to differents objects.")
+            print("The program stops without transferring the series.")
+            return(FALSE)
+        }
     }
-    if (!missing(mp_from_name)) {
-        pos_mp_from <- which(names_saps_from == mp_from_name)
-    }
     
-    # End of the program if the SAP can't be found in one of the workspaces
-    # ie. an unexisting SAP was specified
-    if (all(pos_mp_from == 0)) {
-        warning("The chosen SAP couldn't be found in the first workspace.")
-        return(FALSE)
-    }
     
-    if (all(pos_mp_to == 0) || RJDemetra::count(ws_to) == 0) {
-        print("The chosen SAP doesn't exist in the second workspace.")
+    saps_to <- RJDemetra::get_all_objects(ws_to)
+    names_saps_to <- names(saps_to)
+    
+    # WS to
+    
+    # First case : empty ws
+    if (RJDemetra::count(ws_to) == 0) {
+        print("The ws_to is empty.")
         if (create) {
-            if (missing(mp_to_name)) {
+            if (missing(name_mp_to)) {
                 warning("No name for mp_to has been specified.")
                 print("The program stops without transferring the series.")
                 return(FALSE)
             } else {
-                print("It will be created.")
+                print("A new SAP in ws_to will be created.")
                 RJDemetra::new_multiprocessing(
-                    workspace = ws_to, name = mp_to_name)
+                    workspace = ws_to, name = name_mp_to)
                 pos_mp_to <- RJDemetra::count(ws_to)
             }
         } else {
-            warning("The program stops without transferring the series.")
+            print("The program stops without transferring the series.")
+            return(FALSE)
+        }
+    } else if (missing(name_mp_to) && missing(pos_mp_to)) {
+        pos_mp_to <- 1L
+        name_mp_to <- names_saps_to[1L]
+        print(paste0("No sap has been selected, the first one (", name_mp_to, ") will be used."))
+    } else if (missing(name_mp_to)) {
+        if (pos_mp_to <= RJDemetra::get_all_objects(ws_to) && pos_mp_to > 0L) {
+            name_mp_to <- names_saps_to[pos_mp_to]
+        } else {
+            warning(paste0("There is no SAP n°", pos_mp_to))
+            print("The program stops without transferring the series.")
+            return(FALSE)
+        }
+    } else if (missing(pos_mp_to)) {
+        if (name_mp_to %in% names_saps_to) {
+            pos_mp_to <- which(names_saps_to == name_mp_to)
+        } else if (create) {
+            print(paste0("There is no SAP named ", name_mp_to, " in ws_to."))
+            print("A new SAP in ws_to will be created.")
+            RJDemetra::new_multiprocessing(
+                workspace = ws_to, name = name_mp_to)
+            pos_mp_to <- RJDemetra::count(ws_to)
+        } else {
+            print(paste0("There is no SAP named ", name_mp_to, " in ws_to."))
+            print("The program stops without transferring the series.")
+            return(FALSE)
+        }
+    } else {
+        if (!name_mp_to %in% names_saps_to) {
+            if (create) {
+                print(paste0("There is no SAP named ", name_mp_to, " in ws_to."))
+                print("A new SAP in ws_to will be created.")
+                RJDemetra::new_multiprocessing(
+                    workspace = ws_to, name = name_mp_to)
+                pos_mp_to <- RJDemetra::count(ws_to)
+            } else {
+                print(paste0("There is no SAP named ", name_mp_to, " in ws_to."))
+                print("The program stops without transferring the series.")
+                return(FALSE)
+            }
+        } else if (pos_mp_to > RJDemetra::get_all_objects(ws_to) || pos_mp_to <= 0L) {
+            warning(paste0("There is no SAP n°", pos_mp_to))
+            print("The program stops without transferring the series.")
+            return(FALSE)
+        } else if (names_saps_to[pos_mp_to] != name_mp_to) {
+            warning("The arguments pos_mp_to and name_mp_to are refering to differents objects.")
+            print("The program stops without transferring the series.")
             return(FALSE)
         }
     }
+    
     
     if (print_indications) {
         cat(paste0("Position of the first WS's SAP : ", pos_mp_from, 
@@ -97,8 +170,8 @@ transfer_series <- function(
                    "Position of the second WS's SAP : ", pos_mp_to, "\n"))
     }
     
-    sap_to <- RJDemetra::get_object(ws_to, pos_mp_to)
     sap_from <- RJDemetra::get_object(ws_from, pos_mp_from)
+    sap_to <- RJDemetra::get_object(ws_to, pos_mp_to)
     
     # If a corresponding SAP is found in both workspaces, verification that they are both non empty
     if (RJDemetra::count(sap_from) == 0) {
@@ -108,18 +181,24 @@ transfer_series <- function(
     
     # If all is well:
     # Retrieving the SAPs series...
-    # RJDemetra::count(saps_to[[pos_mp_to]])
-    # RJDemetra::count(saps_from[[pos_mp_from]])
     
-    series_saps_to <- RJDemetra::get_all_objects(sap_to)
     series_saps_from <- RJDemetra::get_all_objects(sap_from)
+    series_saps_to <- RJDemetra::get_all_objects(sap_to)
     
     # ... and their names
-    names_series_to <- names(series_saps_to)
     names_series_from <- names(series_saps_from)
+    names_series_to <- names(series_saps_to)
     
-    # The list of series to transfer: the series in ws_from that are not in ws_to
-    selected_series <- setdiff(names_series_from, names_series_to)
+    # Check de selected series
+    if (missing(selected_series)) {
+        selected_series <- names_series_from
+    } else if (any(!selected_series %in% names_series_from)) {
+        warning("The series ", selected_series[!selected_series %in% names_series_from], " are not in the SAP ", names_sap_from, "from ws_from")
+    }
+    
+    if (!replace_series) {
+        selected_series <- setdiff(selected_series, names_series_to)
+    }
     
     # Replacement of all series specified in the "selected_series" vector
     for (index in seq_along(selected_series)) {
@@ -127,7 +206,7 @@ transfer_series <- function(
         series_name <- selected_series[index]
         position <- which(names_series_from == series_name)
         if (print_indications) {
-            print(paste0("Series n°", index, 
+            cat(paste0("Series n°", index, 
                          ", name: ", series_name, 
                          ", position: ", position))
         }
@@ -135,10 +214,17 @@ transfer_series <- function(
         # The "up-to-date" series version
         extracted_sa_item <- RJDemetra::get_object(sap_from, position)
         
-        # Replacement of the series by its updated version (change made in the reference workspace)
-        add_new_sa_item(sap_to, extracted_sa_item)
+        # Cas de remplacement
+        if (series_name %in% names_series_to) {
+            cat(" - to add...")
+            add_new_sa_item(sap_to, extracted_sa_item)
+        } else {
+            cat(" - to replace...")
+            position <- which(names_series_to == series_name)
+            replace_sa_item(mp_to, pos = position, sa_item = extracted_sa_item)
+        }
         
-        print("Successful transfer")
+        cat(" Successful transfer!\n")
     }
     
     return(TRUE)