+++ NEWS.md
@@ -2,7 +2,7 @@
 
 ## Changes :
 
-* add new function to update_path of **workspaces**, **SAProcessing** and specifically **Sa-items** for csv, xls and xlsx files
+* add new function to `update_path` of **workspaces**, **SAProcessing** and specifically **Sa-items** for csv, xls and xlsx files
 
 
 # rjdworkspace 1.0.0 - Initial release.